Raise gRPC version floor to 1.6 and use new context bridge API
