--- instrumentation/grpc-1.5/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/grpc/v1_5/GrpcContextInstrumentation.java
@@ -1,49 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.grpc.v1_5;
-
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isStatic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
-
-import io.grpc.Context;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import java.util.Collections;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-public class GrpcContextInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("io.grpc.Context");
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return Collections.singletonMap(
-        isMethod()
-            .and(isStatic())
-            .and(named("storage"))
-            .and(returns(named("io.grpc.Context$Storage"))),
-        GrpcContextInstrumentation.class.getName() + "$ContextBridgeAdvice");
-  }
-
-  public static class ContextBridgeAdvice {
-    @Advice.OnMethodEnter(skipOn = Advice.OnDefaultValue.class)
-    public static Object onEnter() {
-      return null;
-    }
-
-    @Advice.OnMethodExit
-    public static void onExit(@Advice.Return(readOnly = false) Context.Storage storage) {
-      storage = GrpcSingletons.STORAGE;
-    }
-  }
-}
