--- instrumentation/grpc-1.5/library/grpc-1.5-library.gradle
@@ -1,15 +1,0 @@
-apply from: "$rootDir/gradle/instrumentation-library.gradle"
-
-def grpcVersion = '1.5.0'
-
-dependencies {
-  library group: 'io.grpc', name: 'grpc-core', version: grpcVersion
-
-  testLibrary group: 'io.grpc', name: 'grpc-netty', version: grpcVersion
-  testLibrary group: 'io.grpc', name: 'grpc-protobuf', version: grpcVersion
-  testLibrary group: 'io.grpc', name: 'grpc-services', version: grpcVersion
-  testLibrary group: 'io.grpc', name: 'grpc-stub', version: grpcVersion
-
-  testImplementation deps.assertj
-  testImplementation project(':instrumentation:grpc-1.5:testing')
-}
