--- instrumentation/grpc-1.5/library/src/test/groovy/io/opentelemetry/instrumentation/grpc/v1_5/GrpcStreamingTest.groovy
@@ -1,22 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.grpc.v1_5
-
-import io.grpc.ManagedChannelBuilder
-import io.grpc.ServerBuilder
-import io.opentelemetry.instrumentation.test.LibraryTestTrait
-
-class GrpcStreamingTest extends AbstractGrpcStreamingTest implements LibraryTestTrait {
-  @Override
-  ServerBuilder configureServer(ServerBuilder server) {
-    return server.intercept(GrpcTracing.create(getOpenTelemetry()).newServerInterceptor())
-  }
-
-  @Override
-  ManagedChannelBuilder configureClient(ManagedChannelBuilder client) {
-    return client.intercept(GrpcTracing.create(getOpenTelemetry()).newClientInterceptor())
-  }
-}
