+++ instrumentation/grpc-1.6/library/src/main/java/io/opentelemetry/instrumentation/grpc/v1_6/GrpcExtractAdapter.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.grpc.v1_6;
+
+import io.grpc.Metadata;
+import io.opentelemetry.context.propagation.TextMapGetter;
+
+final class GrpcExtractAdapter implements TextMapGetter<Metadata> {
+
+  static final GrpcExtractAdapter GETTER = new GrpcExtractAdapter();
+
+  @Override
+  public Iterable<String> keys(Metadata metadata) {
+    return metadata.keys();
+  }
+
+  @Override
+  public String get(Metadata carrier, String key) {
+    return carrier.get(Metadata.Key.of(key, Metadata.ASCII_STRING_MARSHALLER));
+  }
+}
