+++ instrumentation/grpc-1.6/library/src/main/java/io/opentelemetry/instrumentation/grpc/v1_6/GrpcTracingBuilder.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.grpc.v1_6;
+
+import io.opentelemetry.api.OpenTelemetry;
+
+/** A builder of {@link GrpcTracing}. */
+public final class GrpcTracingBuilder {
+
+  private final OpenTelemetry openTelemetry;
+
+  private boolean captureExperimentalSpanAttributes;
+
+  GrpcTracingBuilder(OpenTelemetry openTelemetry) {
+    this.openTelemetry = openTelemetry;
+  }
+
+  /**
+   * Sets whether experimental attributes should be set to spans. These attributes may be changed or
+   * removed in the future, so only enable this if you know you do not require attributes filled by
+   * this instrumentation to be stable across versions
+   */
+  public GrpcTracingBuilder setCaptureExperimentalSpanAttributes(
+      boolean captureExperimentalSpanAttributes) {
+    this.captureExperimentalSpanAttributes = captureExperimentalSpanAttributes;
+    return this;
+  }
+
+  /** Returns a new {@link GrpcTracing} with the settings of this {@link GrpcTracingBuilder}. */
+  public GrpcTracing build() {
+    return new GrpcTracing(openTelemetry, captureExperimentalSpanAttributes);
+  }
+}
