+++ instrumentation/grpc-1.6/library/src/test/groovy/io/opentelemetry/instrumentation/grpc/v1_6/GrpcTest.groovy
@@ -1,0 +1,22 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.grpc.v1_6
+
+import io.grpc.ManagedChannelBuilder
+import io.grpc.ServerBuilder
+import io.opentelemetry.instrumentation.test.LibraryTestTrait
+
+class GrpcTest extends AbstractGrpcTest implements LibraryTestTrait {
+  @Override
+  ServerBuilder configureServer(ServerBuilder server) {
+    return server.intercept(GrpcTracing.create(getOpenTelemetry()).newServerInterceptor())
+  }
+
+  @Override
+  ManagedChannelBuilder configureClient(ManagedChannelBuilder client) {
+    return client.intercept(GrpcTracing.create(getOpenTelemetry()).newClientInterceptor())
+  }
+}
