+++ instrumentation/grpc-1.6/testing/src/main/proto/helloworld.proto
