+++ settings.gradle
@@ -114,9 +114,9 @@ include ':instrumentation:geode-1.4:javaagent'
 include ':instrumentation:google-http-client-1.19:javaagent'
 include ':instrumentation:grails-3.0:javaagent'
 include ':instrumentation:grizzly-2.0:javaagent'
-include ':instrumentation:grpc-1.5:javaagent'
-include ':instrumentation:grpc-1.5:library'
-include ':instrumentation:grpc-1.5:testing'
+include ':instrumentation:grpc-1.6:javaagent'
+include ':instrumentation:grpc-1.6:library'
+include ':instrumentation:grpc-1.6:testing'
 include ':instrumentation:guava-10.0:javaagent'
 include ':instrumentation:guava-10.0:library'
 include ':instrumentation:gwt-2.0:javaagent'
